+++ backend/library/views.py
@@ -1,8 +1,10 @@
 import logging
+import operator
+from functools import reduce
 
 from common.utils import retrieve_group_items
 from django.apps import apps
-from django.db.models import Prefetch
+from django.db.models import Prefetch, Q
 from library_sample_shared.views import LibrarySampleBaseViewSet
 from rest_framework import viewsets
 from rest_framework.response import Response
@@ -21,9 +23,30 @@ logger = logging.getLogger("db")
 
 
 class LibrarySampleTree(viewsets.ViewSet):
-    def get_queryset(self, showAll=True):
+    def filter_and_search(self, queryset, search_string, status_filter):
+        """Helper function for both get_queryset and list action"""
+        if search_string:
+            search_fields = [
+                "name__icontains",
+                "barcode__icontains",
+                "request__name__icontains",
+            ]
+            search_filters = [Q(**{field: search_string}) for field in search_fields]
+            queryset = queryset.filter(reduce(operator.or_, search_filters))
+
+        if status_filter:
+            queryset = queryset.filter(status=int(status_filter))
+
+        return queryset
+
+    def get_queryset(self, show_all=True, search_string=None, status_filter=None):
         libraries_qs = Library.objects.all().only("sequencing_depth")
         samples_qs = Sample.objects.all().only("sequencing_depth")
+
+        libraries_qs = self.filter_and_search(
+            libraries_qs, search_string, status_filter
+        )
+        samples_qs = self.filter_and_search(samples_qs, search_string, status_filter)
 
         queryset = (
             Request.objects.filter(archived=False)
@@ -34,7 +57,8 @@ class LibrarySampleTree(viewsets.ViewSet):
             .only("name")
             .order_by("-create_time")
         )
-        if not showAll:
+
+        if not show_all:
             queryset = queryset.filter(sequenced=False)
         if not self.request.user.is_staff:
             if not self.request.user.is_pi:
@@ -46,12 +70,11 @@ class LibrarySampleTree(viewsets.ViewSet):
 
     def list(self, request):
         """Get the list of libraries and samples."""
-        showAll = False
-        if request.GET.get("showAll") == "True":
-            showAll = True
-        queryset = self.get_queryset(showAll)
 
-        request_id = self.request.query_params.get("node", None)
+        show_all = request.query_params.get("showAll") or False
+        search_string = request.query_params.get("searchString")
+        status_filter = request.query_params.get("statusFilter")
+        request_id = request.query_params.get("node", None)
 
         if request_id and request_id != "root":
             libraries_qs = Library.objects.all().select_related(
@@ -67,6 +90,13 @@ class LibrarySampleTree(viewsets.ViewSet):
                 "library_type",
                 "read_length",
                 "organism",
+            )
+
+            libraries_qs = self.filter_and_search(
+                libraries_qs, search_string, status_filter
+            )
+            samples_qs = self.filter_and_search(
+                samples_qs, search_string, status_filter
             )
 
             queryset = (
@@ -102,9 +132,14 @@ class LibrarySampleTree(viewsets.ViewSet):
                     },
                     400,
                 )
+        else:
+            queryset = self.get_queryset(show_all, search_string, status_filter)
+            serializer = RequestParentNodeSerializer(queryset, many=True)
+            filtered_data = [
+                item for item in serializer.data if item["total_records_count"] != 0
+            ]  # omit rows (requests) that would be empty upon expanding (clicking plus sign)
 
-        serializer = RequestParentNodeSerializer(queryset, many=True)
-        return Response({"success": True, "children": serializer.data})
+            return Response({"success": True, "children": filtered_data})
 
 
 class LibraryViewSet(LibrarySampleBaseViewSet):