+++ backend/static/main-hub/app/view/statistics/Sequences.js
@@ -28,7 +28,7 @@ Ext.define("MainHub.view.statistics.Sequences", {
             xtype: "parkoursearchfield",
             store: "SequencesStatistics",
             emptyText: "Search",
-            width: 200,
+            width: 320,
           },
         ],
       },